requirements.txt
# FastAPI 및 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 데이터베이스
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# 벡터 검색
faiss-cpu==1.7.4
numpy==1.24.3

# 임베딩 모델
openai==1.3.7
cohere==4.37
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1

# 파일 처리
PyPDF2==3.0.1
python-docx==1.1.0

# 설정 관리
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Streamlit 프론트엔드
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.3

# 유틸리티
requests==2.31.0
python-dateutil==2.8.2 